[refactor] App: extract root UI layout to independent composable
