+++ app/src/main/java/com/rwmobi/giphytrending/MainActivity.kt
@@ -9,22 +9,10 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.imePadding
-import androidx.compose.foundation.layout.safeDrawingPadding
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Surface
 import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
-import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
 import androidx.compose.material3.windowsizeclass.calculateWindowSizeClass
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import androidx.navigation.compose.rememberNavController
-import com.rwmobi.giphytrending.ui.components.AppBottomNavigationLayout
-import com.rwmobi.giphytrending.ui.components.AppNavigationRailLayout
-import com.rwmobi.giphytrending.ui.theme.GiphyTrendingTheme
+import com.rwmobi.giphytrending.ui.components.GiphyTrendingApp
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -36,46 +24,9 @@ class MainActivity : ComponentActivity() {
         installSplashScreen()
 
         setContent {
-            val windowSizeClass = calculateWindowSizeClass(this)
-            val navController = rememberNavController()
-            val snackbarHostState = remember { SnackbarHostState() }
-
-            GiphyTrendingTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier
-                        .safeDrawingPadding()
-                        .fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background,
-                ) {
-                    // Select a navigation element based on window size.
-                    when (windowSizeClass.widthSizeClass) {
-                        WindowWidthSizeClass.Compact -> {
-                            AppBottomNavigationLayout(
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .imePadding(),
-                                windowSizeClass = windowSizeClass,
-                                navController = navController,
-                                snackbarHostState = snackbarHostState,
-                            )
-                        }
-
-                        WindowWidthSizeClass.Medium, // tablet portrait
-                        WindowWidthSizeClass.Expanded, // phone landscape mode
-                        -> {
-                            AppNavigationRailLayout(
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .imePadding(),
-                                windowSizeClass = windowSizeClass,
-                                navController = navController,
-                                snackbarHostState = snackbarHostState,
-                            )
-                        }
-                    }
-                }
-            }
+            GiphyTrendingApp(
+                windowSizeClass = calculateWindowSizeClass(this),
+            )
         }
     }
 }
